Introduce global scap locks
Tags: #release-engineering-team

Summary: This allows us to globally lock all scap deploys (regardless of project) in the situation where things are broken and we don't want anyone to deploy something.

Test Plan: Looked at it

Reviewers: mmodell, #release-engineering-team, demon

Reviewed By: #release-engineering-team, demon

Subscribers: thcipriani, jenkins

Differential Revision: https://phabricator.wikimedia.org/D416

+++ scap/utils.py
@@ -393,6 +393,9 @@ def lock(filename):
     :param filename: File to lock
     :raises: LockFailedError on failure
     """
+    if os.path.exists('/var/lock/scap-global-lock'):
+        raise LockFailedError('Failed to lock scap, global lock in place')
+
     lock_fd = None
     try:
         lock_fd = open(filename, 'w+')
